Use new way of enqueueing for alert type highlight assets
